services/web/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.3
gunicorn==20.1.0
psycopg2-binary==2.9.6
Flask-Login==0.6.3
Flask-WTF==1.2.1
Werkzeug==2.3.7
python-dotenv==1.0.1
Flask-Migrate==4.0.5
